fix: look up bridge history by source tx hash, not txMetaId
selectBridgeHistoryForAccount is keyed by txMetaId, but several surfaces
indexed it with the transaction hash, so the lookup missed and bridge
titles/amounts fell back to raw tx data (wrong token) instead of the quote.
Resolve bridge history through the source-tx-hash-keyed view
(useBridgeHistoryItemBySrcTxHash) via a shared findBridgeHistoryItemBySrcTxHash
helper, and use it in the Activity Details screen, its amount header, and the
per-asset activity list. The main Activity list now reuses the same helper.
