+++ app/components/Views/ActivityDetails/ActivityDetails.tsx
@@ -14,7 +14,11 @@ import {
 import { strings } from '../../../../locales/i18n';
 import { useParams } from '../../../util/navigation/navUtils';
 import { selectTransactionMetadataById } from '../../../selectors/transactionController';
-import { selectBridgeHistoryForAccount } from '../../../selectors/bridgeStatusController';
+// eslint-disable-next-line import-x/no-restricted-paths
+import {
+  useBridgeHistoryItemBySrcTxHash,
+  findBridgeHistoryItemBySrcTxHash,
+} from '../../UI/Bridge/hooks/useBridgeHistoryItemBySrcTxHash';
 import type { RootState } from '../../../reducers';
 import { resolveActivityListItemTitle } from '../../UI/ActivityListItemRow/ActivityListItemRow';
 // eslint-disable-next-line import-x/no-restricted-paths -- TODO(ADR-0020): reuses the confirmations speed-up/cancel modal; route-isolation backlog
@@ -63,8 +67,11 @@ const ActivityDetails = () => {
   const preloadedItem = preloadedRef.current.item;
 
   const item = useActivityDetailsItem(txIdentifier, chainId, preloadedItem);
-  const bridgeHistory = useSelector(selectBridgeHistoryForAccount);
-  const bridgeHistoryItem = item?.hash ? bridgeHistory?.[item.hash] : undefined;
+  const { bridgeHistoryItemsBySrcTxHash } = useBridgeHistoryItemBySrcTxHash();
+  const bridgeHistoryItem = findBridgeHistoryItemBySrcTxHash(
+    bridgeHistoryItemsBySrcTxHash,
+    item?.hash,
+  );
   const title = item
     ? resolveActivityListItemTitle(item, bridgeHistoryItem)
     : strings('activity_details.not_found');
